Support hash: searches. Use @ in place of & for intersections.
hash: searches simply find all indexed files whose hashes begin with a
given string of characters. So, if you wanted all files whose hashes
began with 'a', you'd search hash:a or hash:A (all hashes should be
lowercase).

The @ operator is just a synonym for the intersection operator and is
intended to replace it at some point. It's just there to be friendlier
to the shell, though if your shell also treats @ as something special,
then I'm not entirely sure what else to use. Might add named operators,
like intersect: and union: and subtract: in the future as well.